+++ Ginger Island Mainland Adjustments/Ginger Island Mainland Adjustments/ScheduleManager/DataModels/IslandNorthScheduler.cs
@@ -100,7 +100,7 @@ internal static class IslandNorthScheduler
                     // Child2NPC NPCs don't understand "bed", must send them to the bus stop spouse dropoff.
                     ?? (Globals.IsChildToNPC?.Invoke(explorer) == true ? "1800 BusStop -1 23 3" : "1800 bed"));
 
-                Globals.ModMonitor.DebugLog($"Calculated island north schedule for {explorer.Name}");
+                Globals.ModMonitor.DebugOnlyLog($"Calculated island north schedule for {explorer.Name}");
                 explorer.islandScheduleName.Value = "island";
 
                 ScheduleUtilities.ParseMasterScheduleAdjustedForChild2NPC(explorer, renderedSchedule);