+++ Ginger Island Mainland Adjustments/Ginger Island Mainland Adjustments/ScheduleManager/MidDayScheduleEditor.cs
@@ -84,7 +84,7 @@ internal static class MidDayScheduleEditor
         }
         if (npc.IsInvisible)
         {
-            Globals.ModMonitor.DebugLog($"NPC {npc.Name} is invisible, not altering schedule", LogLevel.Trace);
+            Globals.ModMonitor.DebugOnlyLog($"NPC {npc.Name} is invisible, not altering schedule", LogLevel.Trace);
             return false;
         }
         if (npc.Schedule is null)
@@ -100,7 +100,7 @@ internal static class MidDayScheduleEditor
         keys.Sort();
         if (keys.Count == 0 || keys[^1] != GIEndTime)
         {
-            Globals.ModMonitor.DebugLog($"Recieved {npc.Name} to adjust but last schedule key is not {GIEndTime}");
+            Globals.ModMonitor.DebugOnlyLog($"Recieved {npc.Name} to adjust but last schedule key is not {GIEndTime}");
             return false;
         }
         string? schedule = ScheduleUtilities.FindProperGISchedule(npc, SDate.Now());